Move logger.ts into logging/index.ts
