+++ src/compose/network.ts
@@ -3,7 +3,7 @@ import type dockerode from 'dockerode';
 
 import { docker } from '../lib/docker-utils';
 import logTypes = require('../lib/log-types');
-import * as logger from '../logger';
+import * as logger from '../logging';
 import log from '../lib/supervisor-console';
 import * as ComposeUtils from './utils';
