+++ test/integration/compose/volume.spec.ts
@@ -3,7 +3,7 @@ import type { SinonStub } from 'sinon';
 import { stub } from 'sinon';
 import { Volume } from '~/src/compose/volume';
 import * as logTypes from '~/lib/log-types';
-import * as logger from '~/src/logger';
+import * as logger from '~/src/logging';
 
 import Docker from 'dockerode';
