+++ test/integration/device-config.spec.ts
@@ -7,7 +7,7 @@ import { expect } from 'chai';
 
 import * as deviceConfig from '~/src/device-config';
 import * as fsUtils from '~/lib/fs-utils';
-import * as logger from '~/src/logger';
+import * as logger from '~/src/logging';
 import { Extlinux } from '~/src/config/backends/extlinux';
 import { ConfigTxt } from '~/src/config/backends/config-txt';
 import { Odmdata } from '~/src/config/backends/odmdata';
